requirements.txt
# Web Framework
fastapi==0.115.6  # Latest stable - includes starlette security fixes
uvicorn[standard]==0.34.0
python-multipart==0.0.18  # Updated to fix CVE-2024-24762, CVE-2024-53981

# AWS SDK
boto3==1.29.7
botocore==1.32.7

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Rate Limiting
slowapi==0.1.9

# Monitoring
prometheus-fastapi-instrumentator==7.0.0

# Utilities
python-dotenv==1.0.0

# Structured Logging
structlog>=24.0.0
python-json-logger>=2.0.0

# Tracing - OpenTelemetry with AWS X-Ray
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-exporter-otlp-proto-grpc==1.28.0
opentelemetry-instrumentation-fastapi==0.49b0
opentelemetry-instrumentation-boto3sqs==0.49b0
opentelemetry-instrumentation-botocore==0.49b0
opentelemetry-sdk-extension-aws==2.0.2
